\name{makeProject-package}
\alias{makeProject-package}
\docType{package}
\title{Make Project LCFD Framework}
\description{This package was inspired by a stackoverflow.com comment made by Josh Reich.  This package creates an empty framework of files and directories for the "Load, Clean, Func, Do" structure described by Josh.  The concept is simple.  The "main.R" script is called by the user.  It sets global variables and calls all the other scipts necessary.}
\details{
\tabular{ll}{
Package: \tab makeProject\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-05-17\cr
License: \tab GPL-3\cr
}
}
\author{
Noah Silverman 

Maintainer: Noah Silverman <noahsilverman@ucla.edu>
}
\keyword{ package }

